refactor: renamed card content to card group
