+++ app/src/main/java/br/com/useblu/oceands/client/ui/cardgroup/CardGroupActivity.kt
@@ -1,4 +1,4 @@
-package br.com.useblu.oceands.client.ui.cardcontent
+package br.com.useblu.oceands.client.ui.cardgroup
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
@@ -6,21 +6,19 @@ import android.widget.Toast
 import androidx.databinding.DataBindingUtil
 import androidx.lifecycle.ViewModelProvider
 import br.com.useblu.oceands.client.R
-import br.com.useblu.oceands.client.databinding.ActivityCardContentBinding
-import br.com.useblu.oceands.client.databinding.ActivityTransactionListItemBinding
-import br.com.useblu.oceands.client.ui.transactionlistitem.TransactionListItemViewModel
+import br.com.useblu.oceands.client.databinding.ActivityCardGroupBinding
 
-class CardContentActivity : AppCompatActivity() {
+class CardGroupActivity : AppCompatActivity() {
 
-    private lateinit var binding: ActivityCardContentBinding
-    private lateinit var viewModel: CardContentViewModel
+    private lateinit var binding: ActivityCardGroupBinding
+    private lateinit var viewModel: CardGroupViewModel
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        binding = DataBindingUtil.setContentView(this, R.layout.activity_card_content)
+        binding = DataBindingUtil.setContentView(this, R.layout.activity_card_group)
         binding.lifecycleOwner = this
 
-        viewModel = ViewModelProvider(this)[CardContentViewModel::class.java]
+        viewModel = ViewModelProvider(this)[CardGroupViewModel::class.java]
         binding.viewmodel = viewModel
 
         initObservables()
