+++ ocean-components/src/main/res/layout/ocean_card_group.xml
